move validation pytest to test_validate_turtle_file_syntax.py
